+++ src/components/categoriesTable/CategoriesHeader.tsx
